Relax validation of UK postal codes
To be consistent with the stripe.js Card Element,
the only requirement for UK addresses is that
it is non-empty. Remove UK postal code regex.

Rename `ShippingPostalCodeValidator` to `PostalCodeValidator`

Fixes #1988
